Sort inserts consistently to avoid deadlocks
The Postgres docs and various blog posts advise that taking locks
(either implicit, such as with ShareLock, or explicitly, which we're not
doing) needs to be done in a consistent order to avoid deadlocks that
we're seeing.  Since the ShareLock is implicit on the rows we're
inserting, the only way to do consistent lock taking is to order those
inserts, so for all `insert_*` functions, sort the `changes_list`, so
that it is consist and importantly between `insert_addresses` and
`update_balances`, use the same ordering for Addresses.  This eliminates
the need to serialize writes, so BlockImporter is removed.