Don't allow Sequence to be started with a mix of ranges and infinite start
We either use Sequence with a fixed set of ranges in finite mode when
getting the missing blocks down to genesis or in infinite mode when
using it for the realtime index.  We can't completely separate the
support for queuing ranges for the realtime index as that handles
retries, but we can simplify the validation considerably by disallowing
ranges and first to `start_link`.