Split batches on 504 Gateway Timeout
Under the assumption that gateway timeouts occur because it is taking
too long for Sokol nodes to respond to batch requests, 504 Gateway
Timeout should be treated the same as 413 Request Entity Too Large: the
batch should be split in half and retried.

If we still get 504 Gateway Timeouts after this change, we'll need to
add a delay before retry.