Don't select forked transactions where there are no blocks
There was a bug with `where_forked`, where if there are no blocks it will
select **ALL** transactions because no `WHERE` conditions are added to the
`SELECT ... FROM transactions`. Instead of starting the initial query
that way, start with `WHERE false`, so it will select nothing until the
`OR`s for the block hashes and numbers are added.