Don't replace_all fields of internal_transactions on_conflict
Fixes #978

Replace all field except the primary key and timestamps.  For the
primary key fields (transaction_hash, index), leave them as is as they
are the conflict target and have no changed by definition.  For the
timestamps that the least inserted_at and the greatest updated_at to
preserve the bounds of the timeline.