Import stages
Stages will be between Import and the Import.Runner.  Unlike Runners,
which use 1 common Ecto.Multi, a Stage can produce 1 or more independent
Ecto.Multi that will run in separate transactions.  This allows
addresses to be chunked into separate transactions, which releases the
locks between transactions unlike the INSERT-level chunking that was all
in 1 transaction.