Don't generate wei^2 from Wei.mult
Wei.mult required both parameters to be Wei.t, but that meant it was
actually yielding `wei^2` units.  Fix the units by requiring the first
parameter to be wei and the second to be a unitless integer, which makes
more sense for the `gas_price * gas_used` usage as we expect that to be
measured in wei and not wei^2.